# Add Int8 output support to the optimised binary convolution kernels.