--- a/larq_compute_engine/tflite/tests/bconv2d_test.cc
+++ b/larq_compute_engine/tflite/tests/bconv2d_test.cc
@@ -409,8 +409,10 @@ void test_lce_op_output(const std::vector<std::int8_t>& lce_output_data,
     unrounded_builtin_output[i] = unrounded_int8_output;
   }
 
-  EXPECT_THAT(lce_output_data, ::testing::Pointwise(::testing::FloatNear(0.55),
-                                                    unrounded_builtin_output));
+  const float tolerance = high_error_tolerance ? 0.7 : 0.55;
+  EXPECT_THAT(lce_output_data,
+              ::testing::Pointwise(::testing::FloatNear(tolerance),
+                                   unrounded_builtin_output));
 }
 
 template <typename TOutput>