## What do these changes do?

This PR adds support for writing Int8 output to the Aarch64 and Arm32 optimised assembly binary convolution kernels.

## How Has This Been Tested?

CI. The 'big' kernel tests pass locally for Aarch64 and Arm32.

## Benchmark Results

It turns out that benchmarking this change is quite difficult. Usually we use the benchmarking tool and measure the overall latencies of the QuickNet models. However, it appears (verified through profiling) that an Int8-quantised version of QuickNet is substantially slower than a float version because the full-precision ops are slower, in particular the depthwise convolution at the start and the add ops.

Therefore, to benchmark this change I have used the benchmark tool with the Ruy profiler enabled and multiplied the % of time spent on our `BConv2D` op by the average overall latency (`num_runs=250`). This is not a nice way of measuring things - having the profiler enabled *severely* increases run variation - but it should be enough to get a general idea.

These benchmarks were taken on my Android phone (a OnePlus 6), running the QuickNet models single-threaded. The reported values (in ms) are the estimated time spent executing the bconv op.

| Model         | Baseline (float) | Baseline (int8) | PR (int8) |
|---------------|------------------|-----------------|-----------|
| QuickNet      | 11.06            | 190.23          | 10.92     |
| QuickNetLarge | 15.60            | 260.65          | 16.11     |
| QuickNetXL    | 31.73            | 540.93          | 32.86     |

## Related issue number

N/A.