@AdamHillier 2020-09-04
I've updated the PR description with benchmark results (such that they are). I think this is now ready from my end.